+++ CHANGES.md
@@ -1,5 +1,10 @@
 # aws-api 0.8
 
+## DEV
+
+* improved support for apigatewaymanagementapi `:PostToConnection`
+  * see [https://github.com/cognitect-labs/aws-api/#posttoconnection](https://github.com/cognitect-labs/aws-api/#posttoconnection)
+
 ## 0.8.289 / 2019-03-29
 
 * fix signing bug introduced in 0.8.283